+++ docs/CHANGELOG.rst
@@ -23,6 +23,7 @@ Changed
 Added
 -----
 - Add filtering ``Variant`` by ``VariantExperiment`` and ``VariantCall``
+- Add BAM output option to ``vc-gatk4-hc`` 
 
 Changed
 -------