+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 -----
 - Add filtering ``Variant`` by ``VariantExperiment`` and ``VariantCall``
 - Use generic permission filters when filtering variants by related models
+- Add BAM output option to ``vc-gatk4-hc`` 
 
 
 ===================