+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Add filtering ``Variant`` by ``VariantExperiment`` and ``VariantCall``
 - Use generic permission filters when filtering variants by related models
 - Add ``--bam-output`` input argument to ``vc-gatk4-hc``
+- Add ``--max-mnp-distance`` input argument to ``vc-gatk4-hc``
 
 
 ===================