+++ docs/CHANGELOG.rst
@@ -16,6 +16,8 @@ Added
 - Add filtering ``Variant`` by ``VariantExperiment`` and ``VariantCall``
 - Use generic permission filters when filtering variants by related models
 - Return only distinct ``VariantExperiment`` objects
+- Add ``--bam-output`` input argument to ``vc-gatk4-hc``
+- Add ``--max-mnp-distance`` input argument to ``vc-gatk4-hc``
 
 
 ===================